feat(seal): add seals
