+++ src/event/event_data/interaction.rs
@@ -1,4 +1,4 @@
-use super::super::sections::DelegatingEventSeal;
+use super::super::sections::seal::*;
 use crate::error::Error;
 use crate::prefix::SelfAddressingPrefix;
 use crate::state::{EventSemantics, IdentifierState};
@@ -9,11 +9,11 @@ pub struct InteractionEvent {
     #[serde(rename = "prev")]
     pub previous_event_hash: SelfAddressingPrefix,
 
-    pub data: Vec<DelegatingEventSeal>,
+    pub data: Vec<Seal>,
 }
 
 impl EventSemantics for InteractionEvent {
     fn apply_to(&self, state: IdentifierState) -> Result<IdentifierState, Error> {
-        IdentifierState { ..state }
+        Ok(IdentifierState { ..state })
     }
 }
